Reconfigured the initialisation to be done using python script
It needed to be done this way because the passwords are hashed.
Now we can create test users with a variable for the password, which the script then substitutes with the hashed password.
Since this is done via a direct connection to the database, the references for the view-creation files, have shifted to be run as seperate queries inside the db_initialisation